// frontend/app/[lang]/types/generated-strapi.ts
export interface StrapiMedia {
  id: number;
  documentId: string;
  name: string;
  alternativeText?: string | null;
  caption?: string | null;
  width: number;
  height: number;
  formats?: {
    [key: string]: {
      name: string;
      hash: string;
      ext: string;
      mime: string;
      path: string | null;
      width: number;
      height: number;
      size: number;
      sizeInBytes: number;
      url: string;
    };
  } | null;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl?: string | null;
  provider: string;
  provider_metadata?: any | null;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
}

export type StrapiRelation<T> = {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
} & T;

export type StrapiCollection<T> = Array<{
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
} & T>;
// ...
export interface LinksLink {
  url: string;
  newTab?: boolean;
  text: string;
}

export interface LinksSocialLink {
  url: string;
  newTab?: boolean;
  text: string;
  social?: string;
}
// ...
export interface SharedSeo {
  metaTitle: string;
  metaDescription: string;
  shareImage?: StrapiMedia;
}

export interface SharedSlider {
  files?: StrapiMedia[];
}

export interface SharedVideoEmbed {
  url: string;
}

export interface SkillsSkills {
  skill?: string;
}

export interface Article {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  title?: string;
  description: string;
  slug?: string;
  cover?: StrapiMedia;
  category?: StrapiRelation<Category>;
  blocks?: any[];
  authorsBio?: StrapiRelation<Author>;
  seo?: SharedSeo;
}

export interface Author {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  name?: string;
  avatar?: StrapiMedia;
  email?: string;
  articles?: StrapiCollection<Article>;
  jobs?: StrapiCollection<Job>;
}

export interface Category {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  name?: string;
  slug?: string;
  articles?: StrapiCollection<Article>;
  description?: string;
  jobs?: StrapiCollection<Job>;
}

export interface Global {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  metadata?: MetaMetadata;
  hero_section?: ElementsHeroSection;
  favicon: StrapiMedia;
  notificationBanner?: ElementsNotificationBanner;
  navbar?: LayoutNavbar;
  footer?: LayoutFooter;
  cv?: StrapiMedia;
}

export interface Job {
  id: number;
  documentId: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  jobTitle: string;
  description: any;
  fullDescription: any;
  start: string;
  finish?: string;
  fullTime?: boolean;
  company?: string;
  slug?: string;
  cover?: StrapiMedia;
  icon?: StrapiMedia;
  category?: StrapiRelation<Category>;
  authorsBio?: StrapiRelation<Author>;
  blocks?: any[];
  seo?: SharedSeo;
}